+++ server.js
@@ -5,8 +5,8 @@ const {resolve} = require('path');
 const PrettyError = require('pretty-error');
 const finalHandler = require('finalhandler');
 const session = require('express-session');
-const containerRoutes = require('./containerRoutes');
-const userRoutes = require('./userRoutes');
+const containerRoutes = require('./routes/containerRoutes');
+const userRoutes = require('./routes/userRoutes');
 const models = require('./models');
 const Project = models.Project;
 const db = models.db;
